Cambios

+++ purchaseTracking/Models/eTALENT/VACACIONES_DISPONIBLES.cs
@@ -11,8 +11,8 @@ namespace purchaseTracking.Models.eTALENT
         public string NOMBRE { get; set; }
         public string FECHA_INGRESO { get; set; }
         public int TOTAL_VACACIONES { get; set; }
-        public int DIAS_TOMADOS { get; set; }
+        public double DIAS_TOMADOS { get; set; }
         public string CODIGO_EXTERNO { get; set; }
-        public int DIAS_PENDIENTES { get; set; }
+        public double DIAS_PENDIENTES { get; set; }
     }
 }
